Create tokens-and-tokenisation
This is a new page under Concepts. Note that a link to PII information on line 14 will need to be added later.

In the Confluence draft we suggested having the sub-heading "How Meeco utilise tokens", however we didn't feel we had enough knowledge and information to create text. Perhaps this is something we add later?